refactor: semantic changes and refactoring
